Agent class with updated example
