+++ examples/ql_maze_policy.py
@@ -2,12 +2,11 @@ import pygame as pg
 from environments.maze import MazeEnvironment
 from vicero.algorithms.qlearning import Qlearning
 from vicero.policy import RandomPolicy
+from vicero.agent import Agent
 import numpy as np
 
 # In this example the focus is to see frozen policies 
 # compared to each other in real time.
-
-np.random.seed()
 
 board = [[0 ,0 ,0 ,0 ,10,0 ,0 ,0 ],
          [0 ,0 ,-1,-1,-1,0 ,0 ,0 ],
@@ -18,8 +17,6 @@ board = [[0 ,0 ,0 ,0 ,10,0 ,0 ,0 ],
          [0 ,0 ,0 ,-1,-1,0 ,0 ,0 ],
          [0 ,0 ,0 ,1 ,0 ,0 ,0 ,0 ]]
 
-# TODO: move visualization etc. to the MazeEnvironment class..
-
 cell_size  = 48 # the size of one game cell, in pixels
 framerate  = 10 # frames per second
 
@@ -28,19 +25,7 @@ pg.init()
 screen = pg.display.set_mode((cell_size * len(board[0]), cell_size * len(board)))
 clock = pg.time.Clock()
 
-class GameInstance:
-    def __init__(self, env, policy):
-        self.env = env
-        self.policy = policy
-
-    def game_step(self):
-        action = self.policy(self.env.state)
-        _, _, done, _ = self.env.step(action)
-        self.env.draw(screen)
-        if done: self.env.reset()
-
 env = MazeEnvironment(board, cell_size)
-
 ql = Qlearning(env, len(board) ** 2, len(MazeEnvironment.action_space), epsilon=0.15, discretize=lambda state : state[1] * env.size + state[0])
 
 # Random for benchmarking
@@ -58,7 +43,7 @@ policy_two = ql.copy_target_policy()
 ql.train(20000)
 policy_three = ql.copy_target_policy()
 
-game = GameInstance(env, random_policy)
+agent = Agent(env, random_policy)
 
 policies = [('the random policy',          random_policy),
             ('the bad policy',             policy_one),
@@ -69,9 +54,10 @@ while True:
     for policy in policies:
         print('switching to {}'.format(policy[0]))
         
-        game.policy = policy[1]
+        agent.policy = policy[1]
 
         for _ in range(30):
-            game.game_step()
+            agent.step()
+            env.draw(screen)
             pg.display.flip()
             clock.tick(framerate)
